Fixes way to retrieve last transaction

+++ Awarean.BrayaOrtega.RinhaBackend.Q124/Models/ExecuteTransactionResponse.cs
@@ -1,9 +1,9 @@
 namespace Awarean.BrayaOrtega.RinhaBackend.Q124;
 
-public readonly partial struct ExecuteTransactionResponse(bool isSuccess, long limite, long saldo, TransactionExecutionError? error = null)
+public readonly partial struct ExecuteTransactionResponse(bool isSuccess, int limite, int saldo, TransactionExecutionError? error = null)
 {
-    public long Limite { get; } = limite;
-    public long Saldo { get; } = saldo;
+    public int Limite { get; } = limite;
+    public int Saldo { get; } = saldo;
     public bool IsSuccess { get; } = isSuccess;
     public TransactionExecutionError? Error { get; } = error;
 }
